+++ PyrusServiceDeskIOS/PyrusServiceDesk/Navigation/PyrusServiceDeskController.swift
@@ -155,7 +155,7 @@ class PyrusServiceDeskController: PSDNavigationController, PSDMainController {
         if self.parent == nil {
             self.dismiss(animated: animated, completion: {
                 PyrusServiceDesk.stopCallback?.onStop()
-                PyrusLogger.shared.logEvent("Chat closed ")
+                PyrusLogger.shared.logEvent(" Chat closed ")
                 PyrusServiceDeskController.clean()
                 PyrusServiceDesk.isStarted = false
             })
@@ -165,7 +165,7 @@ class PyrusServiceDeskController: PSDNavigationController, PSDMainController {
             let fake = self.parent
             fake?.dismiss(animated: animated, completion: {
                 PyrusServiceDesk.stopCallback?.onStop()
-                PyrusLogger.shared.logEvent("Chat closed ")
+                PyrusLogger.shared.logEvent(" Chat closed ")
                 PyrusServiceDeskController.clean()
                 PyrusServiceDesk.isStarted = false
             })
